Rename NIO to Network
